BLink class
stores the link between a Model Object and a graphics object, and the lastposition of that BObject. When the BObject is moved, that last position does not get updated. By comparing o.getXY to xy, we can determine if the object has moved/has to be updated in the view. this functionality lives in the hasMoved function.